Cleanup. Add turn_to_regular
