+++ accounts-db/src/account_info.rs
@@ -10,11 +10,6 @@ use {
         is_zero_lamport::IsZeroLamport,
     },
     modular_bitfield::prelude::*,
-    solana_pubkey::Pubkey,
-    std::{
-        collections::HashMap,
-        sync::atomic::{AtomicU64, Ordering},
-    },
 };
 
 /// offset within an append vec to account data
@@ -176,227 +171,9 @@ impl AccountInfo {
     }
 }
 
-#[derive(Debug)]
-pub struct IrregularEntry {
-    pub foo: u64,
-    pub bar: u32,
-    pub baz: u64,
-}
-
-impl IrregularEntry {
-    pub fn from_regular(regular: RegularEntry) -> Self {
-        Self {
-            foo: regular.file_id() as u64,
-            bar: regular.offset_reduced() as u32,
-            baz: 1,
-        }
-    }
-}
-
-#[bitfield(bits = 64)]
-#[repr(C)]
-#[derive(Debug, Default, Copy, Clone, Eq, PartialEq)]
-pub struct RegularEntry {
-    pub common_info: CommonEntryInfo,
-    /// Bits available for future use
-    pub reserved: B9,
-    /// Storage id, capped at 2^20-1, since maximum number of storages is 1000000
-    pub file_id: B20,
-    pub offset_reduced: B31,
-    pub is_zero_lamports: bool,
-}
-
-#[bitfield(bits = 3)]
-#[derive(Debug, Default, Copy, Clone, Eq, PartialEq, BitfieldSpecifier)]
-pub struct CommonEntryInfo {
-    pub is_regular: bool,
-    pub reserved: B2,
-}
-
-#[derive(Debug)]
-pub enum FixedKindIndexEntry<'a> {
-    Regular(RegularEntry),
-    Irregular(&'a IrregularEntry),
-}
-
-#[bitfield(bits = 64)]
-#[repr(C)]
-#[derive(Debug, Default, Copy, Clone, Eq, PartialEq)]
-pub struct IrregularEntryUnsafeAddress {
-    pub common_info: CommonEntryInfo,
-    /// A non-zero heap memory address holding `IrregularEntry` (i.e. 8-byte aligned struct) down-shifted by 3 bits
-    pub shifted_address: B61,
-}
-
-pub struct CompactEntryHolder {
-    raw_atomic: AtomicU64,
-}
-
-impl CompactEntryHolder {
-    pub fn from_regular(regular: RegularEntry) -> Self {
-        Self::from_unsafe_holder(UnsafeEntryHolder::from_regular(regular))
-    }
-
-    pub fn from_irregular(irregular: IrregularEntry) -> Self {
-        Self::from_unsafe_holder(UnsafeEntryHolder::from_irregular(irregular))
-    }
-
-    pub fn as_fixed_entry(&self) -> FixedKindIndexEntry {
-        self.make_unsafe_holder().into_fixed_entry()
-    }
-
-    pub fn turn_to_irregular(&self) -> FixedKindIndexEntry {
-        let mut current_raw = self.raw_atomic.load(Ordering::Acquire);
-        loop {
-            let current_holder = UnsafeEntryHolder::from_raw(current_raw);
-            match current_holder.into_fixed_entry() {
-                FixedKindIndexEntry::Regular(regular) => {
-                    let new_holder =
-                        UnsafeEntryHolder::from_irregular(IrregularEntry::from_regular(regular));
-                    match self.raw_atomic.compare_exchange(
-                        current_raw,
-                        unsafe { new_holder.raw },
-                        Ordering::SeqCst,
-                        Ordering::SeqCst,
-                    ) {
-                        Ok(_) => return new_holder.into_fixed_entry(),
-                        Err(different_raw) => {
-                            // entry that we created and tried to store will not be used, so deallocate it
-                            new_holder.deallocate_irregular();
-                            // continue loop to re-check if the entry is still regular
-                            current_raw = different_raw
-                        }
-                    }
-                }
-                e @ FixedKindIndexEntry::Irregular(_) => return e,
-            }
-        }
-    }
-
-    fn make_unsafe_holder(&self) -> UnsafeEntryHolder {
-        UnsafeEntryHolder::from_raw(self.raw_atomic.load(Ordering::Relaxed))
-    }
-
-    fn from_unsafe_holder(holder: UnsafeEntryHolder) -> Self {
-        Self {
-            raw_atomic: AtomicU64::new(unsafe { holder.raw }),
-        }
-    }
-}
-
-impl Drop for CompactEntryHolder {
-    fn drop(&mut self) {
-        self.make_unsafe_holder().deallocate_irregular();
-    }
-}
-
-union UnsafeEntryHolder {
-    raw: u64,
-    regular: RegularEntry,
-    irregular: IrregularEntryUnsafeAddress,
-}
-
-impl UnsafeEntryHolder {
-    const ALIGNED_PTR_SHIFT: u64 = 3;
-
-    pub fn from_raw(raw: u64) -> Self {
-        Self { raw }
-    }
-
-    pub fn from_regular(regular: RegularEntry) -> Self {
-        assert!(regular.common_info().is_regular());
-        Self { regular }
-    }
-
-    pub fn from_irregular(irregular: IrregularEntry) -> Self {
-        let irregular_ptr = Box::into_raw(Box::new(irregular));
-        let irregular = IrregularEntryUnsafeAddress::new()
-            .with_common_info(CommonEntryInfo::new().with_is_regular(false))
-            .with_shifted_address(Self::make_ptr_shifted_address(irregular_ptr));
-        Self { irregular }
-    }
-
-    pub fn is_regular(&self) -> bool {
-        unsafe { self.regular.common_info().is_regular() }
-    }
-
-    pub fn into_fixed_entry<'a>(self) -> FixedKindIndexEntry<'a> {
-        unsafe {
-            if self.is_regular() {
-                FixedKindIndexEntry::Regular(self.regular)
-            } else {
-                let ptr = self.get_irregular_ptr();
-                FixedKindIndexEntry::Irregular(ptr.as_ref().unwrap())
-            }
-        }
-    }
-
-    fn deallocate_irregular(self) {
-        if !self.is_regular() {
-            unsafe {
-                let ptr = self.get_irregular_mut_ptr();
-                drop(Box::from_raw(ptr));
-            }
-        }
-    }
-
-    fn make_ptr_shifted_address(ptr: *mut IrregularEntry) -> u64 {
-        (ptr as u64) >> Self::ALIGNED_PTR_SHIFT
-    }
-
-    fn get_irregular_ptr(&self) -> *const IrregularEntry {
-        unsafe {
-            assert!(!self.is_regular());
-            let addr = self.irregular.shifted_address() << Self::ALIGNED_PTR_SHIFT;
-            addr as *const IrregularEntry
-        }
-    }
-
-    fn get_irregular_mut_ptr(&self) -> *mut IrregularEntry {
-        unsafe {
-            assert!(!self.is_regular());
-            let addr = self.irregular.shifted_address() << Self::ALIGNED_PTR_SHIFT;
-            addr as *mut IrregularEntry
-        }
-    }
-}
-
-#[derive(Default)]
-pub struct EntryMap {
-    pub map: HashMap<Pubkey, CompactEntryHolder>,
-}
-
 #[cfg(test)]
 mod test {
     use {super::*, crate::append_vec::MAXIMUM_APPEND_VEC_FILE_SIZE};
-
-    #[test]
-    fn test_new() {
-        let regular = RegularEntry::new()
-            .with_common_info(CommonEntryInfo::new().with_is_regular(true))
-            .with_file_id(4534)
-            .with_offset_reduced(99)
-            .with_is_zero_lamports(true);
-        let uni_e = CompactEntryHolder::from_regular(regular);
-
-        let irregular = IrregularEntry {
-            foo: 42,
-            bar: 45,
-            baz: 13,
-        };
-        let _uni_p = CompactEntryHolder::from_irregular(irregular);
-        // assert_eq!(
-        //     format!(
-        //         "e={:?} p={:?}",
-        //         uni_e.as_fixed_entry(),
-        //         uni_p.as_fixed_entry()
-        //     ),
-        //     ""
-        // );
-        assert_eq!(format!("t={:?}", uni_e.turn_to_irregular()), "");
-
-        let _map = EntryMap::default();
-    }
 
     #[test]
     fn test_limits() {
