+++ accounts-db/src/accounts_index.rs
@@ -1,4 +1,5 @@
 mod account_map_entry;
+pub mod caterpillar_map;
 pub(crate) mod in_mem_accounts_index;
 mod iter;
 mod roots_tracker;
